Update fatbackbone_from_general_curvature.m
added support for the new from-system orientation option

I haven't tested this because I haven't been able to get a system that uses a general curvature to work properly, and I don't think it's worth my time to figure that out. But the code is almost identical to the tested fatbackbone_from_curvature_bases.m so I think it's pretty safe to say it will work.

If not, it definitely won't break the old functionality.