Polished animation
Animating is now in its own function. It uses softspace and repeats the
gait three times. For now, the number of gait cycles and the frames per
cycle are hard-coded.

To collect the accumulated displacement from past gait cycles, I'm
currently just adding the displacement. I think it should be a matrix
transformation instead.